+++ src/Cache/PackageMetadata.php
@@ -2,12 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Cpx;
+namespace Cpx\Cache;
+
+use Cpx\Packages\Package;
 
 class PackageMetadata
 {
     public function __construct(
-        public Package $package,
+        public readonly Package $package,
         public ?string $lastUpdatedAt = null,
         public ?string $lastRunAt = null,
     ) {}
